Adds GET to create_saved_list_item
Background: The page shows Method \"GET\" not allowed when navigating and posting through the api/list-items-create path.

I think the reason is that a get request is made when navigating to this path. A workaround is to allow for get requests. Additionally, it would be nice to have feedback once the posted data is saved, which line 60 will do: `return Response(serializer.data).